+++ src/App.js
@@ -5,23 +5,17 @@ import {
   FormControl,
   InputLabel,
   MenuItem,
-  Paper,
   Select,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
   TextField,
   Typography,
 } from "@mui/material";
 import SearchIcon from "@mui/icons-material/Search";
 import "./App.css";
 import { useState } from "react";
-import { Triangle } from "react-loader-spinner";
 import { CalculateRouteCost, FindAllPaths } from "./RouteFinder";
-import Toast from "./Components/toast";
+import Toast from "./Components/Toast";
+import TableComponent from "./Components/Table";
+import LoadingComponent from "./Components/Loading";
 
 function App() {
   const [formData, setFormData] = useState("");
@@ -58,53 +52,6 @@ function App() {
       });
     }
     setLoading(false);
-  };
-
-  const LoadingComponent = () => {
-    return (
-      <Box
-        sx={{
-          display: "flex",
-          flexDirection: "column",
-          backgroundColor: "#f5f5f5",
-          minHeight: 300,
-          justifyContent: "center",
-          alignItems: "center",
-        }}
-      >
-        <Triangle color="#06BE8B" height={100} width={100} />
-        <Typography sx={{ color: "#06BE8B" }}>Searching Routes....</Typography>
-      </Box>
-    );
-  };
-
-  const CustomTableComponent = () => {
-    return (
-      <TableContainer component={Paper}>
-        <Table sx={{ minWidth: 650 }} aria-label="simple table">
-          <TableHead>
-            <TableRow>
-              <TableCell colSpan={2}>Routes</TableCell>
-              <TableCell />
-              <TableCell>{routes.cost ? "Cost" : "Path Count"}</TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            <TableRow>
-              {typeof routes === "object" && (
-                <>
-                  <TableCell colSpan={2}>{routes.route}</TableCell>
-                  <TableCell />
-                  <TableCell>
-                    {routes.cost ? routes.cost : routes.totalPaths}
-                  </TableCell>
-                </>
-              )}
-            </TableRow>
-          </TableBody>
-        </Table>
-      </TableContainer>
-    );
   };
   return (
     <div className="App">
@@ -171,7 +118,7 @@ function App() {
           marginTop: 5,
         }}
       >
-        {loading ? <LoadingComponent /> : <CustomTableComponent />}
+        {loading ? <LoadingComponent /> : <TableComponent routes={routes} />}
       </Box>
     </div>
   );
